refactor: HandlerAdapter의 구현체 클래스 네이밍 변경
