+++ mvc/src/main/java/nextstep/mvc/controller/tobe/HandlerExecutionHandlerAdapter.java
@@ -1,24 +1,21 @@
-package com.techcourse;
+package nextstep.mvc.controller.tobe;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 
-public class ManualHandlerAdapter implements HandlerAdapter {
+public class HandlerExecutionHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(final Object handler) {
-        return handler instanceof Controller;
+        return handler instanceof HandlerExecution;
     }
 
     @Override
     public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response,
                                final Object handler) throws Exception {
-        Controller controller = (Controller) handler;
-        String viewName = controller.execute(request, response);
-        return new ModelAndView(new JspView(viewName));
+        HandlerExecution handlerExecution = (HandlerExecution) handler;
+        return handlerExecution.handle(request, response);
     }
 }
